Logging improvements
Feedzilla was giving me an error exception without traceback. By
changing logging.error to log.exception the traceback is also logged and
way more "debugable".

Finally the error was an invalid feed XML that IMHO should be
treated as a warning and not as a error message, so I've added an
special treatment for that.